requirements.txt
deep-translator==1.11.4
translators==5.9.1
huggingface-hub==0.22.2
manga-ocr==0.1.11
numpy==1.24.2
opencv-python==4.9.0.80
pillow==10.3.0
ultralytics==8.1.43
sentencepiece==0.2.0
torch==2.5.0
paddlepaddle
paddleocr
requests>=2.31.0
gradio>=5.0.0
